requirements.txt
request
colorama
beautifulsoup4